apps/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
python-dotenv==1.0.1
anthropic==0.34.0
openai==1.51.0
google-generativeai==0.8.0
tenacity==8.3.0
structlog==24.4.0